Move AVX2 ISA into submodule in preparation for adding AVX-512 ISA
